downloader/requirements.txt
python-dotenv
pymongo
selenium
minio
